\name{k.prior}
\alias{k.prior}

\title{ simple prior }
\description{
  This function is used by the metropolis sampler
}
\usage{
k.prior(seg, ps)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{seg}{  }
  \item{ps}{  }
}
\details{
 
}
\value{

}
\references{  }
\author{}
\note{}
\seealso{  }
\examples{

}
\keyword{manip }
