\name{pimg.list}
\alias{pimg.list}
\alias{pimg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Create a collection of probability images, for MCMC binning. }
\description{
 Pimage lists.
}
\usage{
pimg(xmin, xmax, xn, ymin, ymax, yn)
pimg.list(times, xlim, ylim, img.dim, Z = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xmin}{  \code{xmin}  }
  \item{xmax}{  \code{xmax}  }
  \item{xn}{  \code{xn}  }
  \item{ymin}{  \code{ymin}  }
  \item{ymax}{  \code{ymax}  }
  \item{yn}{  \code{yn}  }



  \item{times}{  \code{times}  }
  \item{xlim}{  \code{xlim}  }
  \item{ylim}{  \code{ylim}  }
  \item{img.dim}{  \code{img.dim}  }
  \item{Z}{  \code{Z}  }
}
\value{
 returns a Pimage list
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}
