\name{julday}
\alias{julday}
\alias{julcent}
\title{ Julian day and Julian century calculations from date-time values }
\description{
  [definition and rationale of Julian]
}
\usage{
julday(tm)

julcent(time)
}
\arguments{
  \item{tm}{ vector of date-times }

  \item{time}{ vector of date-times }
}
\details{
}
\value{
}
\references{[Astro and NOAA]}
\author{ Michael D. Sumner }
\note{ 
}
\seealso{ }
\examples{
}
\keyword{ manip }