\name{julday}
\alias{julday}
\alias{julcent}
\title{ Julian day and Julian century calculations from date-time values }
\description{
  Date values required by \code{\link{solar}}.
}
\usage{
julday(tm)

julcent(time)
}
\arguments{
  \item{tm}{ vector of date-times }

  \item{time}{ vector of date-times }
}
\value{
  return numeric values
}
\references{\url{https://gml.noaa.gov/grad/solcalc/azel.html}}
\author{ Michael D. Sumner }

\keyword{ manip }
